backend/sliceManager/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.24,<1.0
kafka-python>=2.0.2,<3.0
pydantic>=2.5,<3.0